feat: 코드 형식 확인 인증

+++ android/festago/domain/src/main/java/com/festago/festago/repository/SchoolRepository.kt
@@ -1,5 +1,5 @@
 package com.festago.festago.repository
 
 interface SchoolRepository {
-    suspend fun loadSchoolEmail(schoolId: Int): Result<String>
+    suspend fun loadSchoolEmail(schoolId: Long): Result<String>
 }
